rag-chatbot-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
qdrant-client==1.8.0
psycopg2-binary
pydantic==2.5.0
openai==1.3.7
pyjwt==2.8.0
python-multipart==0.0.6
requests==2.31.0
numpy==1.24.3
tiktoken==0.5.1
SQLAlchemy==2.0.23